Corrected regex
I found the regex pattern was omitting fields that did not have any quotes in them (ex. destip). I have adjusted and tested and this pattern seems to be working.

I also added the newline='' option to the "with open..." statement so that the outputted CSV does not have alternating blank rows.